// src/API.ts
export type CreateNoteInput = {
  id?: string | null,
  name: string,
  description?: string | null,
};
// ...
export type Note = {
  __typename: "Note",
  id: string,
  name: string,
  description?: string | null,
  createdAt: string,
  updatedAt: string,
};

export type UpdateNoteInput = {
  id: string,
  name?: string | null,
  description?: string | null,
};

export type DeleteNoteInput = {
  id: string,
};
// ...
export type ModelNoteConnection = {
  __typename: "ModelNoteConnection",
  items:  Array<Note >,
  nextToken?: string | null,
};
// ...
export type ListNotesQuery = {
  listNotes?:  {
    __typename: "ModelNoteConnection",
    items:  Array< {
      __typename: "Note",
      id: string,
      name: string,
      description?: string | null,
      createdAt: string,
      updatedAt: string,
    } >,
    nextToken?: string | null,
  } | null,
};